+++ earth2studio/lexicon/__init__.py
@@ -27,8 +27,12 @@ from .goes import GOESLexicon
 from .hrrr import HRRRFXLexicon, HRRRLexicon
 from .isd import ISDLexicon
 from .jpss import JPSSATMSLexicon, JPSSCrISLexicon, JPSSLexicon
-from .metop import MetOpAMSUALexicon, MetOpAVHRRLexicon, MetOpMHSLexicon
-from .metop_iasi import MetOpIASILexicon
+from .metop import (
+    MetOpAMSUALexicon,
+    MetOpAVHRRLexicon,
+    MetOpIASILexicon,
+    MetOpMHSLexicon,
+)
 from .mrms import MRMSLexicon
 from .ncar import NCAR_ERA5Lexicon
 from .planetary_computer import (
